
\name{twoStageGwasPower-internal}
\alias{[.unit}
\alias{[.unit.arithmetic}
\alias{[.unit.list}
\alias{.GRID.STATE}
\alias{.grid.gpar.names}
\alias{.grid.loaded}
\alias{.grid.unit.list}
\alias{Ops.unit}
\alias{Summary.unit}
\alias{.unit}
\alias{condnorm}
\alias{find.p.joint.alpha}
\alias{find.p.joint}
\alias{find.p.joint.alt}
\alias{freqConvert}
\alias{genotypingSavings}
\alias{p.joint.fun.alt}
\alias{p.joint.fun}
\alias{print.twoStageGwasPower}
\alias{print.twoStageNull}
\alias{termsInt.alt}
\alias{termsInt}
\alias{twoStagePower}

\title{Internal twoStageGwasPower Functions}
\description{
 Internal twoStageGwasPower functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
