\name{twoStageGwasPower-package}
\alias{twoStageGwasPower-package}

\docType{package}
\title{
Power and critical values for a two-stage genome-wide association study
}
\description{
Computes the critical values for a two-stage gwas association study and
compute power for replication and joint analyses.
}
\details{
\tabular{ll}{
Package: \tab twoStageGwasPower\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2012-08-28\cr
License: \tab gpl (>=2)\cr
}

}
\author{
Dirk F Moore <dirkfmoore@gmail.com>
}
\references{
Skol AD, Scott, LJ, Abecasis GR, Boehnke M (2006) Nature Genetics
               doi:10.1038/ng1706 \cr
  http://www.sph.umich.edu/csg/abecasis/CaTS/
}

\keyword{gwas}
